fix(release): resume tagged draft publication
